copilotkit-pydantic/first-party-mcp-servers/jira/requirements.txt
# Jira MCP Server Dependencies

# Atlassian Python API - Jira client
atlassian-python-api>=3.41.0

# FastMCP - MCP server framework
fastmcp>=0.1.0

# Pydantic for type validation
pydantic>=2.0.0

# Caching
cachetools>=5.3.0

# HTTP server for FastMCP
uvicorn[standard]>=0.27.0

